More API changes and added Main runner implementation
This provides a basic runner that can execute a ScriptFactory to provide a script.  While this may not solve all the needs, it may reduce the needs to make a main class to get the results from a test.  I expect we will want to expand (and unit test) this later.