Fix for exceptions during parallel running tests
This fixes exceptions which may occur after a test has finished or stopped, but other tasks are still running.  Specific examples would be after a failure and futures were canceled (and then we attempted to shutdown the pool).  Or for progress steps which don't have futures to represent them.

This should be cleaned up, but this at least gets the functionality to be correct.

I also cleaned up the build scripts and updated threadly