Tweaks around script builder lifecycle and other things
This makes the builder lifecycle more specific.  Making sure we Futures we return are accurate and resolve when multiple instances complete.
This requires the user to manually copy the chain before reusing it.  In the future we should make the automatic.
There are additional tweaks and improvements in memory usage.