This resolves #7 and resolves #8
The failure to stop a test during a parallel chain of execution has been fixed.  More implementation moved into the abstract "ScriptCollectionRunner" and that all now shares a single implementation (that should be correct).
I also added the ability to adjust the script execution rate.  This is done at the time of building the script, and is able to run at different rates at different points in time.