Fix for #11
We now copy the ExecutionAssistant before any modifications are made to it.
That copy is passed down the step chain (until another change occurs to it).
That way we can control what changes could be passed up the step chain, some we still want to (for example stopping execution due to step failure).