Execute every ExecutionItem through the ExecutionAssistant
This further reduces the jar and memory footprint, and in some ways is simpler.
Instead of having each item determine how to execute the item (ie ScriptStepRunner executing through ExecutionAssistant and determining if it should block or not...or how parallel vs sequential steps are executed, and how limits are applied).  Now the ExecutionAssistant figures out how an item should be executed.
This means that the ExecutionAssistant can determine if this needs to be put through the limiter, or if it can be run in thread (or forced to be async via a provided boolean).  Basically we must provide it the information needed so it can execute them the right way.
Then callers can decide if they must block on the returned future or not, or how else they want to control the execution flow.
This solves the need for the wrapper in the ParallelScriptBuilder in another way (one which I think is cleaner and gets more advantage towards reducing memory usage and wrapper classes).