Add ability to balance execution chains without rate limiting
This can be very useful for replicating production load levels on different endpoints.
Previously you had manually set a rate limit on chains which completed faster.  This required a lot of manual tuning, and was not able to re-balance the chains with unexpected or variable execution durations.

Unfortanetly this results in a bunch of blocked threads, rather than just queued tasks, as well as a chunk of additional complexity.  But in time I hope we can improve that.