Changes to how we build the scripts and handle parameters
This extracts out the parameter handling from ScriptFactory into "ParameterStore".
This change allows us to make AbstractScriptFactoryInitializer also extend "ParameterStore".  That way not only do script factories have the ability to set parameters, but also extending "ScriptRunner"'s may set their own parameters.
For example in ambushGUI we may want to have parameters to set the dimensions of the gui window.
I considered removing AbstractScriptFactoryInitializer and just put it all into ScriptRunner.  However I don't like the idea that it is expected the only one building scripts factories is doing so from a main function.  Instead I made AbstractScriptFactoryInitializer a bit more generic so that this assumption is less clear.  That does result in a tiny amount of duplicated code in ambushGUI for handling the error conditions where we want to exit the app.