Add ability for "maintenance" steps
These steps can run synchroniously, or asynchronously.  In either case, they provide an ability to run a step which is only used for managing the runtime of the script.
These steps wont be included in the final statistics, and in the async case wont even fail the test if they throw an exception (though any set ExceptionHandler will be notified of such failures).